Move validation function into separate module
