+++ datameta/api/ui/convert.py
@@ -23,19 +23,11 @@ import csv
 import pandas as pd
 
 from ... import security, samplesheet, errors
+from ...models import MetaDatum, MetaDataSet, MetaDatumRecord
+from ...utils import formatted_mrec_value_str
 from ..metadata import get_all_metadata
 
 log = logging.getLogger(__name__)
-
-
-def formatted_mrec_value(value, datetimefmt):
-    if datetimefmt is not None:
-        try:
-            return datetime.datetime.fromisoformat(value).strftime(datetimefmt)
-        except ValueError:
-            pass
-    return value
-
 
 def get_samplesheet_reader(file_like_obj):
     """Given a file with tabular data which is either in delimited plain text
@@ -92,7 +84,7 @@ def convert_samplesheet(db, file_like_obj, filename, user):
         # Datetimes are converted according to the format string. Empty strings
         # are converted to None aka null, i.e. setting metadata to an empty
         # string value is not possible through the convert API.
-        return [ { mdname : None if not row[mdname] else formatted_mrec_value(row[mdname], metadata_datetimefmt[mdname]) for mdname in metadata_names } for _, row in submitted_metadata.iterrows() ]
+        return [ { mdname : None if not row[mdname] else formatted_mrec_value_str(row[mdname], metadata_datetimefmt[mdname]) for mdname in metadata_names } for _, row in submitted_metadata.iterrows() ]
     except Exception as e:
         log.error(e)
         raise samplesheet.SampleSheetReadError("Unknown error")
